Changed ContentType to a string reference
ContentType was previously a strongly typed reference to an in-memory
instance of a ContentType object. This is convenient for the Editor and
the Save() methods, which require the ContentType's properties to
determine how Attributes should be treated. Aside from these scenarios,
however, this is not only not required, but introduces a lot of
dependencies (i.e., a Topic cannot create children, as required by
SetRelationships(), and to create a new Topic, there must be a reference
to a cached copy of the topic graph). By migrating this to a string
reference, this dependency is removed.