Renamed `Attribute` to `AttributeDescriptor`
This update should remain backward compatible with current projects and databases, although client code should also be updated to use `AttributeDescriptor` as soon as possible, as this is the preferrent nomenclature. The term `Attribute` is ambiguous between the `AttributeValue` and `AttributeDescriptor`. This is especially true since "Attribute" is so ubiquitously used throughout the library for `AttributeValue` (e.g., `Topic.Attributes`).