Changed `IsCurrentRelationship()` to `GetRelationship()`
This doesn't actually reduce the lines of code, but it does make the lines of code more readable by putting more of the logic in the local function. The only reason it doesn't reduce the lines of code is because we're placing each function call on multiple lines for readability. This does help reduce the amount of complex `(… && …) || (… && …)` style expressions.