Established `TopicViewModelLookupService`
If an application is simply using the OOTB view models, then the `TopicViewModelLookupService` provides a convenient wrapper for `StaticTypeLookupService` which automatically adds the out-of-the-box view models. In addition, it can be derived by applications to add any application-specific view models. If an application adds a view model of the same name as an out-of-the-box version, that view model will be used instead of the factory default. Updated the `FakeViewModelLookupService` as an example. Updated documentation.