Implemented alternative to `Ensures()`
Due to the removal of Code Contracts, the `Ensures()` method no longer serves a function. In many cases, the `Ensures()` method was only intended to satisfy Code Contracts analysis module. In some cases, however, there is a legitimate chance that the `Ensures()` condition might not be met—such as a factory method failing to create an instance of an object. In these cases, a `NullReferenceException` is thrown if the return object is null, typically using the `return [object]?? throw new NullReferenceException(…)` construction. This provides a functional alternative to `Ensures()`.