Updated `GitVersionTask` to 4.0.0
As a matter of housekeeping, but also in preparation for an eventual conversion to .NET core, I updated the packages to use the latest `GitVersionTask` build. There have been some updates to better support the .NET Core file format. This may impact the build process, but shouldn't impact the functionality itself.