Allow `FullSemVer` for *NuGet*
If we were publishing (pre-releases) to *NuGet*, we wouldn't be permitted to use the full semantic versioning enabled by the *NuGet 2.0* specification (for reasons that aren't entirely clear to me). Since we are only publishing pre-released to *Azure DevOps*, however, we _should_ be able to take advantage of these, which provides more granularity over builds off individual branches.