Updated `UtilPack`
Older versions of `UtilPack`—including, notably, the version included with `GitVersion`—have a limitation that binds them to a specific version of `nuget`. If that version differers from the version installed locally, errors can occur. This is resolved in the latest version—though that version isn't (yet) available in `GitVersion`. To remedy this, the latest package version of `UtilPack` is referenced independently.

See e.g. https://github.com/stazz/UtilPack/issues/34#issuecomment-438805936