Removed setters for collections
Setters on collections represent a potential risk, as they allow the entire collection to be overwritten, and potentially bind the property to a shared reference. This may be desirable in some circumstances, but should not be the assumed default—as rightly called out by code analysis. To mitigate this, all collection properties have been updated to be self-initializing. This also avoids a scenario where collections may return a null, which is unexpected and unintuitive.