Moved `Contracts` to `Internal` namespace
Established new `Internal` namespace, which will be used for classes that are primarily intended to support the *OnTopic* infrastructure, but not be consumed directly by customers. Importantly, _these may still be marked as public_ so that they can be used across assemblies. The distinction is not that they are literally `internal`, but that they are not intended for use by customers, and thus their interfaces and functionality should not be assumed to be stable across versions.