Handled initial `CA1062` validation warnings
The `CA1062` warning occurs when parameters or variables are not validated prior to use. This usually means that a guard clause is missing—through it can sometimes mean that they should actually be marked as optional, and calls to them bypassed via e.g. a condition or escape clause. This commit provides additional hints for Roslyn's flow analysis to ensure that nullability annotation contexts are understood. In most cases, this just entails adding a `Contract.Requires()` call as a guard clause.