Moved legacy `DefaultConfiguration` members to new `ConfigurableAttributeDescriptor`
In the legacy **ASP.NET WebForms** version of the `Topic-Editor`, a delimited string, `DefaultConfiguration`, contained an untyped assortment of attribute type configuration values, which could be parsed using the `Configuration` dictionary, and accessed with the `GetConfigurationValue()` method. Those aren't needed with the new `AttributeTypeDescriptor` approach, which allow those value to be set and accessed using attributes that are part of the schema. To allow the library to move forward while maintaining backward compatibility, these members are moved off to a new `ConfigurableAttributeDescriptor` class in the `Ignia.Topics.Web` assembly. This is a breaking change, as it will require that all `AttributeTypeDescriptor` objects in the database be updated to a new `ConfigurableAttributeDescriptor` content type. It will also require that the default `TypeLookupService` on `TopicFactory` be replaced with the `WebFormsTopicLookupService` in **WebForms** projects. To help enforce this logic, and force projects to use one of the two type systems, the `AttributeDescriptor` class is now marked as `abstract`.