Established empty `Ignia.Topics.Web.Host` project
Once complete, this will provide a simple test bed for testing the **ASP.NET WebForms** version of the library (i.e., `Ignia.Topics.Web`), while also providing a barebones implementation guide. The current version simply represents an out-of-the-box **Visual Studio** template. We'll be filling it in later. We want to commit this early stage, though, so the updates needed to complete the project are laid out in the commit history.