Introduced default `.gitignore`
The legacy `.gitignore` file was left behind as it would have carried in a lot of nonsensical commits that are not relevant to the newly filtered repository. Instead, a fresh new `.gitignore` is being introduced based on the legacy version. Since the version history for this file isn't especially critical to us, this helps provide a clean history, while maintaining the functionality.